refactor: ErrorMessage, ViewText, WinRewardData 열거형 패키지 이동
