+++ src/main/java/lotto/view/InputView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-import lotto.model.ErrorMessage;
+import lotto.Data.ErrorMessage;
 
 public class InputView {
